#103 extract avfoundation usage into separated module
