+++ permissions-avfoundation/src/iosMain/kotlin/dev/icerock/moko/permissions/avfoundation/AVCaptureDelegate.kt
@@ -1,13 +1,13 @@
 /*
- * Copyright 2022 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2025 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.permissions.ios
+package dev.icerock.moko.permissions.avfoundation
 
 import dev.icerock.moko.permissions.DeniedAlwaysException
 import dev.icerock.moko.permissions.Permission
-import dev.icerock.moko.permissions.PermissionState
 import dev.icerock.moko.permissions.PermissionDelegate
+import dev.icerock.moko.permissions.PermissionState
 import dev.icerock.moko.permissions.mainContinuation
 import platform.AVFoundation.AVAuthorizationStatus
 import platform.AVFoundation.AVAuthorizationStatusAuthorized
@@ -31,7 +31,7 @@ class AVCaptureDelegate(
             AVAuthorizationStatusAuthorized -> return
             AVAuthorizationStatusNotDetermined -> {
                 val isGranted: Boolean = suspendCoroutine { continuation ->
-                    AVCaptureDevice.requestAccess(type) { continuation.resume(it) }
+                    AVCaptureDevice.Companion.requestAccess(type) { continuation.resume(it) }
                 }
                 if (isGranted) return
                 else throw DeniedAlwaysException(permission)
@@ -54,7 +54,7 @@ class AVCaptureDelegate(
     }
 
     private fun currentAuthorizationStatus(): AVAuthorizationStatus {
-        return AVCaptureDevice.authorizationStatusForMediaType(type)
+        return AVCaptureDevice.Companion.authorizationStatusForMediaType(type)
     }
 }
