+++ permissions-camera/src/iosMain/kotlin/dev/icerock/moko/permissions/camera/CameraPermission.ios.kt
@@ -2,7 +2,7 @@ package dev.icerock.moko.permissions.camera
 
 import dev.icerock.moko.permissions.Permission
 import dev.icerock.moko.permissions.PermissionDelegate
-import dev.icerock.moko.permissions.ios.AVCaptureDelegate
+import dev.icerock.moko.permissions.avfoundation.AVCaptureDelegate
 import platform.AVFoundation.AVMediaTypeVideo
 
 actual val cameraDelegate: PermissionDelegate = AVCaptureDelegate(
