+++ permissions-microphone/src/iosMain/kotlin/dev/icerock/moko/permissions/microphone/MicrophonePermission.ios.kt
@@ -2,7 +2,7 @@ package dev.icerock.moko.permissions.microphone
 
 import dev.icerock.moko.permissions.Permission
 import dev.icerock.moko.permissions.PermissionDelegate
-import dev.icerock.moko.permissions.ios.AVCaptureDelegate
+import dev.icerock.moko.permissions.avfoundation.AVCaptureDelegate
 import platform.AVFoundation.AVMediaTypeAudio
 
 actual val recordAudioDelegate: PermissionDelegate = AVCaptureDelegate(
