+++ settings.gradle.kts
@@ -19,6 +19,7 @@ include(":permissions-camera")
 include(":permissions-contacts")
 include(":permissions-gallery")
 include(":permissions-location")
+include(":permissions-avfoundation")
 include(":permissions-microphone")
 include(":permissions-motion")
 include(":permissions-notifications")
